Add files via upload
actualización ,Interfaz, combó box, descripción

+++ src/ps_Inventario/Users.java
@@ -14,6 +14,7 @@ import java.sql.Statement;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.swing.DefaultListModel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import static ps_Inventario.Dashboard.content;
 
@@ -57,16 +58,19 @@ public class Users extends javax.swing.JPanel {
         jLabel5 = new javax.swing.JLabel();
         delete = new javax.swing.JPanel();
         jLabel1 = new javax.swing.JLabel();
+        jLabel6 = new javax.swing.JLabel();
         edit = new javax.swing.JPanel();
         jLabel2 = new javax.swing.JLabel();
+        jLabel8 = new javax.swing.JLabel();
         nuevo = new javax.swing.JPanel();
         jLabel3 = new javax.swing.JLabel();
+        jLabel7 = new javax.swing.JLabel();
         jScrollPane1 = new javax.swing.JScrollPane();
         jTable1 = new javax.swing.JTable();
 
         setBackground(new java.awt.Color(238, 238, 238));
         setMinimumSize(new java.awt.Dimension(750, 430));
-        setPreferredSize(new java.awt.Dimension(750, 430));
+        setPreferredSize(new java.awt.Dimension(890, 430));
         setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
 
         body.setBackground(new java.awt.Color(255, 255, 255));
@@ -77,9 +81,9 @@ public class Users extends javax.swing.JPanel {
         Title.setText("Usuarios");
         add(Title, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 10, -1, -1));
 
-        jSeparator2.setForeground(new java.awt.Color(0, 153, 255));
+        jSeparator2.setForeground(new java.awt.Color(248, 130, 41));
         jSeparator2.setPreferredSize(new java.awt.Dimension(250, 10));
-        add(jSeparator2, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 70, 620, 10));
+        add(jSeparator2, new org.netbeans.lib.awtextra.AbsoluteConstraints(150, 40, 580, 10));
 
         usrnm.setBackground(new java.awt.Color(238, 238, 238));
         usrnm.setForeground(new java.awt.Color(102, 102, 102));
@@ -93,7 +97,7 @@ public class Users extends javax.swing.JPanel {
                 usrnmMouseReleased(evt);
             }
         });
-        add(usrnm, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 40, 620, 30));
+        add(usrnm, new org.netbeans.lib.awtextra.AbsoluteConstraints(150, 10, 580, 30));
 
         search.setBackground(new java.awt.Color(18, 90, 173));
         search.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
@@ -117,10 +121,10 @@ public class Users extends javax.swing.JPanel {
         jLabel4.setText("Buscar");
         search.add(jLabel4, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 5, 60, 20));
 
-        jLabel5.setIcon(new javax.swing.ImageIcon(getClass().getResource("/ps_Inventario/images/Sin título-1.png"))); // NOI18N
-        search.add(jLabel5, new org.netbeans.lib.awtextra.AbsoluteConstraints(1, 5, 30, 20));
+        jLabel5.setIcon(new javax.swing.ImageIcon(getClass().getResource("/ps_Inventario/images/Search-Male-User-5.png"))); // NOI18N
+        search.add(jLabel5, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 30, 30));
 
-        add(search, new org.netbeans.lib.awtextra.AbsoluteConstraints(660, 40, 90, 30));
+        add(search, new org.netbeans.lib.awtextra.AbsoluteConstraints(750, 10, 100, 30));
 
         delete.setBackground(new java.awt.Color(153, 9, 0));
         delete.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
@@ -143,9 +147,12 @@ public class Users extends javax.swing.JPanel {
         jLabel1.setForeground(new java.awt.Color(255, 255, 255));
         jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
         jLabel1.setText("Borrar");
-        delete.add(jLabel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 5, 60, 20));
+        delete.add(jLabel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 5, 60, 20));
 
-        add(delete, new org.netbeans.lib.awtextra.AbsoluteConstraints(670, 400, 80, 30));
+        jLabel6.setIcon(new javax.swing.ImageIcon(getClass().getResource("/ps_Inventario/images/Remove-Male-User-5.png"))); // NOI18N
+        delete.add(jLabel6, new org.netbeans.lib.awtextra.AbsoluteConstraints(2, 5, -1, -1));
+
+        add(delete, new org.netbeans.lib.awtextra.AbsoluteConstraints(660, 400, 90, 30));
 
         edit.setBackground(new java.awt.Color(5, 115, 0));
         edit.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
@@ -167,9 +174,12 @@ public class Users extends javax.swing.JPanel {
         jLabel2.setForeground(new java.awt.Color(255, 255, 255));
         jLabel2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
         jLabel2.setText("Editar");
-        edit.add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 0, 60, 30));
+        edit.add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 5, 60, 20));
 
-        add(edit, new org.netbeans.lib.awtextra.AbsoluteConstraints(580, 400, 80, -1));
+        jLabel8.setIcon(new javax.swing.ImageIcon(getClass().getResource("/ps_Inventario/images/Edit-Male-User-5.png"))); // NOI18N
+        edit.add(jLabel8, new org.netbeans.lib.awtextra.AbsoluteConstraints(2, 5, -1, -1));
+
+        add(edit, new org.netbeans.lib.awtextra.AbsoluteConstraints(560, 400, 90, 30));
 
         nuevo.setBackground(new java.awt.Color(18, 90, 173));
         nuevo.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
@@ -204,9 +214,12 @@ public class Users extends javax.swing.JPanel {
                 jLabel3MousePressed(evt);
             }
         });
-        nuevo.add(jLabel3, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 0, 60, 30));
+        nuevo.add(jLabel3, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 5, 60, 20));
 
-        add(nuevo, new org.netbeans.lib.awtextra.AbsoluteConstraints(490, 400, 80, 30));
+        jLabel7.setIcon(new javax.swing.ImageIcon(getClass().getResource("/ps_Inventario/images/Add-Male-User-5.png"))); // NOI18N
+        nuevo.add(jLabel7, new org.netbeans.lib.awtextra.AbsoluteConstraints(2, 5, -1, -1));
+
+        add(nuevo, new org.netbeans.lib.awtextra.AbsoluteConstraints(460, 400, 90, 30));
 
         jTable1.setBackground(new java.awt.Color(238, 238, 238));
         jTable1.setFont(new java.awt.Font("Segoe UI", 0, 11)); // NOI18N
@@ -260,7 +273,7 @@ public class Users extends javax.swing.JPanel {
             jTable1.getColumnModel().getColumn(0).setPreferredWidth(10);
         }
 
-        add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 90, 750, 300));
+        add(jScrollPane1, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 90, 850, 300));
     }// </editor-fold>//GEN-END:initComponents
 
     private void usrnmMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_usrnmMousePressed
@@ -359,24 +372,37 @@ public class Users extends javax.swing.JPanel {
                     i++;
                 }
                 int id = Integer.parseInt(list[idcell][0]);
+                String name = list[idcell][1];
+                String apellido = list[idcell][2];
+                int dni = Integer.parseInt(list[idcell][3]);
+                
                 if(id <= 0){
                     javax.swing.JOptionPane.showMessageDialog(this, "Debe seleccionar el usuario a borrar. \n", "AVISO", javax.swing.JOptionPane.INFORMATION_MESSAGE);
                 }
                 else{
-                    Statement stm2=null;
-                    try {
-                    stm2 = reg.createStatement();
-                    } catch (SQLException ex) {
-                    Logger.getLogger(Reports.class.getName()).log(Level.SEVERE, null, ex);
+                    int I = JOptionPane.showConfirmDialog(this, "Desea Eliminar el Usuario con el Dni:  \""+dni+"\" \n \""+name+" "+apellido+"\".");
+                    
+                    if(I == 0){
+                        Statement stm2=null;
+                        try {
+                        stm2 = reg.createStatement();
+                        } catch (SQLException ex) {
+                        Logger.getLogger(Reports.class.getName()).log(Level.SEVERE, null, ex);
+                        }
+                        try {
+                        stm2.executeUpdate("DELETE FROM `users` WHERE `id` = "+id+"");
+                        javax.swing.JOptionPane.showMessageDialog(this, "¡Usuario borrado! \n", "HECHO", javax.swing.JOptionPane.INFORMATION_MESSAGE);
+                        GetUsers();
+                        stm2.close();
+                         } catch (SQLException ex) {
+                             Logger.getLogger(Reports.class.getName()).log(Level.SEVERE, null, ex);
+                         }
+                    }else if (I == 1){
+                        javax.swing.JOptionPane.showMessageDialog(this, "¡Accion Cancelada! \n", "HECHO", javax.swing.JOptionPane.INFORMATION_MESSAGE);
                     }
-                    try {
-                    stm2.executeUpdate("DELETE FROM `users` WHERE `id` = "+id+"");
-                    javax.swing.JOptionPane.showMessageDialog(this, "¡Usuario borrado! \n", "HECHO", javax.swing.JOptionPane.INFORMATION_MESSAGE);
-                    GetUsers();
-                    } catch (SQLException ex) {
-                    Logger.getLogger(Reports.class.getName()).log(Level.SEVERE, null, ex);
-                    }
+                    
                 }
+                stm.close();
             }
         } catch (SQLException ex) {
             Logger.getLogger(Users.class.getName()).log(Level.SEVERE, null, ex);
@@ -535,6 +561,9 @@ public class Users extends javax.swing.JPanel {
     private javax.swing.JLabel jLabel3;
     private javax.swing.JLabel jLabel4;
     private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JSeparator jSeparator2;
     private javax.swing.JTable jTable1;
